Depend on httpcomponents plugin, not library
Depend on the Jenkins Apache httpcomponents plugin instead of the library.
Reduces the size of the plugin and assures that the plugin will use the
Apache httpclient that is installed on the Jenkins controller.

Exclude junit and hamcrest core from the json-simple dependency so that
those two transient dependencies are not included in the packaged plugin.
There is no reason for this plugin to deliver a copy of either of those
libraries.